Rename skeleton files
