+++ ui/packages/shared/components/src/index.tsx
@@ -28,9 +28,9 @@ import ParcaContext, {ParcaContextProvider, defaultValue, useParcaContext} from 
 import Pill, {PillVariant} from './Pill';
 import ResponsiveSvg from './ResponsiveSvg';
 import Select, {type SelectElement} from './Select';
-import IcicleGraphSkeleton, {IcicleActionButtonPlaceholder} from './Skeletons/IcicleGraph';
-import MetricsGraphSkeleton from './Skeletons/MetricsGraph';
-import TableSkeleton, {TableActionButtonPlaceholder} from './Skeletons/Table';
+import IcicleGraphSkeleton, {IcicleActionButtonPlaceholder} from './Skeletons/IcicleGraphSkeleton';
+import MetricsGraphSkeleton from './Skeletons/MetricsGraphSkeleton';
+import TableSkeleton, {TableActionButtonPlaceholder} from './Skeletons/TableSkeleton';
 import Spinner from './Spinner';
 import Tab from './Tab';
 import Table from './Table';
